routemitra-fullmvp/backend/main.py
"""
RouteMitra — Full MVP Backend
FastAPI + Redis + PostgreSQL + Twilio (SMS + WhatsApp)

Run:
  pip install -r requirements.txt
  uvicorn main:app --reload --port 8000
"""

import os
import json
import random
from datetime import datetime
from contextlib import asynccontextmanager

from fastapi import FastAPI, Form, Request, HTTPException
from fastapi.responses import PlainTextResponse, JSONResponse
from fastapi.middleware.cors import CORSMiddleware
from dotenv import load_dotenv

load_dotenv()

# Optional Redis (falls back to in-memory dict for demo)
try:
    import redis.asyncio as aioredis
    REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379")
    redis_client = None
    USE_REDIS = True
except ImportError:
    USE_REDIS = False
    redis_client = None

# In-memory fallback for demo (no Redis needed)
BUS_POSITIONS = {}
COIN_BALANCES = {}
CHECKINS = []

from sms_parser import parse_sms
from eta_engine import calculate_eta
from reply_generator import generate_reply
from data.routes import ROUTES


# ---------------------------------------------------------------------------
# Startup / Shutdown
# ---------------------------------------------------------------------------

@asynccontextmanager
async def lifespan(app: FastAPI):
    global redis_client
    if USE_REDIS:
        try:
            redis_client = aioredis.from_url(REDIS_URL, decode_responses=True)
            await redis_client.ping()
            print("✅ Redis connected")
        except Exception:
            print("⚠ Redis not available — using in-memory fallback")
            redis_client = None
    yield
    if redis_client:
        await redis_client.close()


app = FastAPI(
    title="RouteMitra Full MVP API",
    description="AI-powered offline transit — Ludhiana",
    version="2.0.0",
    lifespan=lifespan,
)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_methods=["*"],
    allow_headers=["*"],
)


# ---------------------------------------------------------------------------
# Redis helpers (with in-memory fallback)
# ---------------------------------------------------------------------------

async def redis_set(key: str, value: str, ex: int = 300):
    if redis_client:
        await redis_client.set(key, value, ex=ex)
    else:
        BUS_POSITIONS[key] = value

async def redis_get(key: str) -> str | None:
    if redis_client:
        return await redis_client.get(key)
    return BUS_POSITIONS.get(key)

async def redis_keys(pattern: str) -> list[str]:
    if redis_client:
        return await redis_client.keys(pattern)
    return [k for k in BUS_POSITIONS if k.startswith(pattern.replace("*", ""))]


# ---------------------------------------------------------------------------
# ── ROUTES API ──
# ---------------------------------------------------------------------------

@app.get("/api/routes")
async def list_routes():
    return {
        "city": "Ludhiana",
        "routes": [
            {
                "id": rid,
                "name": info["name"],
                "stops": info["stops"],
                "frequency_min": info["frequency_min"],
            }
            for rid, info in ROUTES.items()
        ]
    }


@app.get("/api/eta/{route_id}")
async def get_eta(route_id: str, stop: str | None = None):
    route_id = route_id.upper()
    if route_id not in ROUTES:
        raise HTTPException(404, f"Route {route_id} not found")

    route = ROUTES[route_id]
    target_stop = stop or route["stops"][0]

    # Check Redis for live crowdsourced position first
    live_key = f"bus:{route_id}:position"
    live_data = await redis_get(live_key)
    if live_data:
        live = json.loads(live_data)
        source = "crowdsourced"
    else:
        live = None
        source = "simulated"

    eta_data = calculate_eta(
        route_id=route_id,
        stops=route["stops"],
        user_stop=target_stop,
        frequency_min=route["frequency_min"],
    )

    return {
        "route_id": route_id,
        "route_name": route["name"],
        "source": source,
        **eta_data,
    }


# ---------------------------------------------------------------------------
# ── LIVE GPS (Crowdsourced) ──
# ---------------------------------------------------------------------------

@app.post("/api/checkin")
async def passenger_checkin(
    route_id: str,
    stop: str,
    user_id: str = "anonymous",
    lat: float = 30.9,
    lng: float = 75.8,
):
    """
    Passenger checks in — shares live location.
    Earns Transit Coins. Updates Redis with bus position.
    """
    route_id = route_id.upper()
    if route_id not in ROUTES:
        raise HTTPException(404, "Route not found")

    # Save live position to Redis
    position_data = {
        "stop": stop,
        "lat": lat,
        "lng": lng,
        "reported_by": user_id,
        "timestamp": datetime.now().isoformat(),
    }
    await redis_set(f"bus:{route_id}:position", json.dumps(position_data), ex=300)

    # Award Transit Coins
    current = int(COIN_BALANCES.get(user_id, 0))
    COIN_BALANCES[user_id] = current + 10
    CHECKINS.append({"user_id": user_id, "route_id": route_id, "stop": stop})

    return {
        "success": True,
        "coins_earned": 10,
        "total_coins": COIN_BALANCES[user_id],
        "message": "Shukriya! 10 Transit Coins mile hain 🪙",
    }


@app.get("/api/coins/{user_id}")
async def get_coins(user_id: str):
    coins = COIN_BALANCES.get(user_id, 0)
    discount = min(50, int(coins / 10))  # Max 50% discount
    return {
        "user_id": user_id,
        "coins": coins,
        "discount_percent": discount,
        "message": f"{coins} coins = {discount}% discount agli ticket par",
    }


# ---------------------------------------------------------------------------
# ── TWILIO SMS WEBHOOK ──
# ---------------------------------------------------------------------------

@app.post("/webhook/sms", response_class=PlainTextResponse)
async def sms_webhook(Body: str = Form(...), From: str = Form(default="unknown")):
    parsed = parse_sms(Body)
    eta_data = None

    if parsed["command"] == "ETA":
        route = ROUTES[parsed["route_id"]]
        stop = parsed["stop"] or route["stops"][0]
        eta_data = calculate_eta(
            route_id=parsed["route_id"],
            stops=route["stops"],
            user_stop=stop,
            frequency_min=route["frequency_min"],
        )

    reply_text = generate_reply(parsed, eta_data)
    return PlainTextResponse(
        content=f'<?xml version="1.0" encoding="UTF-8"?><Response><Message>{reply_text}</Message></Response>',
        media_type="application/xml",
    )


# ---------------------------------------------------------------------------
# ── WHATSAPP BOT WEBHOOK ──
# ---------------------------------------------------------------------------

@app.post("/webhook/whatsapp", response_class=PlainTextResponse)
async def whatsapp_webhook(Body: str = Form(...), From: str = Form(default="unknown")):
    """
    WhatsApp uses same Twilio webhook — but richer responses.
    User sends location → we find nearest bus.
    User sends text → same as SMS bot.
    """
    # Location message format: "Latitude:30.9 Longitude:75.8"
    if "Latitude:" in Body and "Longitude:" in Body:
        reply = (
            "📍 Location mili!\n"
            "Aapke paas ki buses:\n"
            "• Bus 3 (Dugri→Civil) — ~8 min\n"
            "• Bus 5 (Model Town→Jagraon) — ~15 min\n"
            "\nETA BUS 3 likhein exact time ke liye."
        )
    else:
        parsed = parse_sms(Body)
        eta_data = None
        if parsed["command"] == "ETA":
            route = ROUTES[parsed["route_id"]]
            stop = parsed["stop"] or route["stops"][0]
            eta_data = calculate_eta(
                route_id=parsed["route_id"],
                stops=route["stops"],
                user_stop=stop,
                frequency_min=route["frequency_min"],
            )
        reply = generate_reply(parsed, eta_data)

    return PlainTextResponse(
        content=f'<?xml version="1.0" encoding="UTF-8"?><Response><Message>{reply}</Message></Response>',
        media_type="application/xml",
    )


# ---------------------------------------------------------------------------
# ── ADMIN DASHBOARD API ──
# ---------------------------------------------------------------------------

@app.get("/api/admin/heatmap")
async def get_heatmap():
    """Returns stop-wise demand data for admin heatmap."""
    stop_demand = {}
    for checkin in CHECKINS:
        route = ROUTES.get(checkin["route_id"], {})
        stop = checkin["stop"]
        stop_demand[stop] = stop_demand.get(stop, 0) + 1

    # Add simulated baseline data for demo
    baseline = {
        "Ludhiana Railway Station": 142,
        "Clock Tower": 98,
        "Ghumar Mandi": 76,
        "Civil Lines": 110,
        "Model Town": 88,
        "Dugri Phase 2": 65,
        "PAU Gate": 120,
        "Ludhiana Bus Stand": 155,
    }
    for stop, count in baseline.items():
        stop_demand[stop] = stop_demand.get(stop, 0) + count

    return {
        "heatmap": [
            {"stop": stop, "demand": count}
            for stop, count in sorted(stop_demand.items(), key=lambda x: -x[1])
        ],
        "total_checkins": len(CHECKINS),
        "active_users": len(COIN_BALANCES),
    }


@app.get("/api/admin/revenue")
async def get_revenue():
    """Simulated revenue dashboard for bus owners."""
    return {
        "today": {
            "digital_checkins": len(CHECKINS) + random.randint(40, 80),
            "coins_issued": sum(COIN_BALANCES.values()),
            "estimated_revenue_inr": (len(CHECKINS) + 60) * 12,
        },
        "routes_performance": [
            {"route_id": rid, "checkins": random.randint(20, 100), "avg_delay_min": random.randint(2, 15)}
            for rid in ROUTES
        ],
    }


@app.get("/api/health")
async def health():
    redis_ok = False
    if redis_client:
        try:
            await redis_client.ping()
            redis_ok = True
        except Exception:
            pass
    return {
        "status": "ok",
        "service": "RouteMitra Full MVP",
        "redis": "connected" if redis_ok else "fallback (in-memory)",
        "city": "Ludhiana",
    }


if __name__ == "__main__":
    import uvicorn
    print("🚌 RouteMitra Full MVP starting on http://localhost:8000")
    print("   Docs: http://localhost:8000/docs")
    uvicorn.run("main:app", host="0.0.0.0", port=8000, reload=True)
