routemitra-fullmvp/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
twilio==9.3.2
python-dotenv==1.0.1
httpx==0.27.2
redis==5.0.8
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pydantic==2.9.2
